+++ common/ordered_list.h
@@ -1,6 +1,8 @@
 #ifndef ORDEREDLIST_H
 #define ORDEREDLIST_H
 
+#include <stdlib.h>
+
 typedef struct OrderedListNode{
     void* data;
     struct OrderedListNode* next;
@@ -9,7 +11,7 @@ typedef struct OrderedListNode{
 
 OrderedListNode* ordered_list_create(void* data, int(*comparer)(const void*, const void*));
 void ordered_list_insert_into(OrderedListNode** list, void* data);
-const OrderedListNode* ordered_list_get(const OrderedListNode* list, const int index);
+const OrderedListNode* ordered_list_get(const OrderedListNode* list, const size_t index);
 void ordered_list_destroy(OrderedListNode* list);
 OrderedListNode* ordered_list_pop(OrderedListNode** list);
